requirements.txt
streamlit>=1.32.0
anthropic>=0.18.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
openpyxl>=3.1.2
openai>=1.30.0
pillow>=10.3.0
